refactor(BoardFactory): 테스트를 위해 보드 초기화 로직이 map을 반환하도록 리팩토링
